+++ R/feature_importance.R
@@ -15,7 +15,7 @@
 #'   (`perf_metric_name`) and seed (`seed`) are also returned.
 #'
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' results <- run_ml(otu_small, "glmnet", kfold = 2, cv_times = 2)
 #' names(results$trained_model$trainingData)[1] <- "dx"
 #' get_feature_importance(results$trained_model,
