+++ man/get_feature_importance.Rd
@@ -71,7 +71,7 @@ Calculates feature importance using a trained model and test data. Requires
 the \code{future.apply} package.
 }
 \examples{
-\donttest{
+\dontrun{
 results <- run_ml(otu_small, "glmnet", kfold = 2, cv_times = 2)
 names(results$trained_model$trainingData)[1] <- "dx"
 get_feature_importance(results$trained_model,
