+++ man/combine_hp_performance.Rd
@@ -21,7 +21,7 @@ Named list:
 Combine hyperparameter performance metrics for multiple train/test splits generated by, for instance, \href{http://www.schlosslab.org/mikropml/articles/parallel.html}{looping in R} or using a \href{https://github.com/SchlossLab/mikropml-snakemake-workflow}{snakemake workflow} on a high-performance computer.
 }
 \examples{
-\donttest{
+\dontrun{
 results <- lapply(seq(100, 102), function(seed) {
   run_ml(otu_small, "glmnet", seed = seed, cv_times = 2, kfold = 2)
 })
