+++ man/calc_perf_metrics.Rd
@@ -36,7 +36,7 @@ Dataframe of performance metrics.
 Get performance metrics for test data
 }
 \examples{
-\donttest{
+\dontrun{
 results <- run_ml(otu_small, "glmnet", kfold = 2, cv_times = 2)
 calc_perf_metrics(results$test_data,
   results$trained_model,
