### 2020-12-15 b9d5dd7
Create function to remove columns with non-zero/non-NA values in only one row or fewer
