### 2020-12-15 0080f99
Return removed cols from remove_singleton_columns() too
