### 2020-12-15 c4ba344
Incorporate remove_singleton_columns() in preprocess_data()
