+++ tests/testthat/test-preprocess_data.R
@@ -19,7 +19,9 @@ test_df <- data.frame(
 
 test_that("preprocess_data works", {
   expect_equal(
-    expect_message(preprocess_data(test_df, "outcome"), "Removed "),
+    expect_message(preprocess_data(test_df, "outcome",
+                                   prefilter_threshold = -1),
+                   "Removed "),
     list(dat_transformed = structure(list(
       outcome = c(
         "normal", "normal",
@@ -38,7 +40,10 @@ test_that("preprocess_data works", {
       "var2_a"
     )), removed_feats = c("var5", "var6", "var11"))
   )
-  expect_equal(expect_message(preprocess_data(test_df, "outcome", group_neg_corr = FALSE)), list(dat_transformed = structure(list(
+  expect_equal(expect_message(preprocess_data(test_df, "outcome",
+                                              prefilter_threshold = -1,
+                                              group_neg_corr = FALSE)),
+               list(dat_transformed = structure(list(
     outcome = c(
       "normal", "normal",
       "cancer"
@@ -86,7 +91,8 @@ test_that("preprocess_data works", {
     )
   )
   expect_equal(
-    preprocess_data(test_df[1:3, c("outcome", "var4")], "outcome"),
+    preprocess_data(test_df[1:3, c("outcome", "var4")], "outcome",
+                    prefilter_threshold = -1),
     list(
       dat_transformed = dplyr::tibble(
         outcome = c("normal", "normal", "cancer"),
@@ -97,7 +103,9 @@ test_that("preprocess_data works", {
     )
   )
   expect_equal(
-    expect_message(preprocess_data(test_df[1:3, ], "outcome", method = NULL)),
+    expect_message(preprocess_data(test_df[1:3, ], "outcome",
+                                   method = NULL,
+                                   prefilter_threshold = -1)),
     list(dat_transformed = structure(list(outcome = c(
       "normal", "normal",
       "cancer"
@@ -122,7 +130,10 @@ test_that("preprocess_data works", {
   )
   expect_error(preprocess_data(test_df[1:3, c("outcome", "var5")], "outcome"))
   expect_equal(
-    expect_message(preprocess_data(test_df[1:3, ], "outcome", method = c("range"))),
+    expect_message(preprocess_data(test_df[1:3, ],
+                                   "outcome",
+                                   method = c("range"),
+                                   prefilter_threshold = -1)),
     list(dat_transformed = structure(list(outcome = c(
       "normal", "normal",
       "cancer"
@@ -145,7 +156,11 @@ test_that("preprocess_data works", {
       "var6", "var11"
     ))
   )
-  expect_equal(expect_message(preprocess_data(test_df[1:3, ], "outcome", remove_var = "zv")), list(dat_transformed = structure(list(
+  expect_equal(expect_message(preprocess_data(test_df[1:3, ],
+                                              "outcome",
+                                              remove_var = "zv",
+                                              prefilter_threshold = -1)),
+               list(dat_transformed = structure(list(
     outcome = c(
       "normal", "normal",
       "cancer"
@@ -163,7 +178,10 @@ test_that("preprocess_data works", {
     "var2_a"
   )), removed_feats = c("var5", "var6", "var11")))
   expect_equal(expect_message(
-    preprocess_data(test_df[1:3, ], "outcome", remove_var = NULL),
+    preprocess_data(test_df[1:3, ],
+                    "outcome",
+                    remove_var = NULL,
+                    prefilter_threshold = -1),
     "Removing"
   ), list(dat_transformed = structure(list(
     outcome = c(
@@ -186,7 +204,11 @@ test_that("preprocess_data works", {
     "var6", "var11"
   )))
   expect_equal(
-    expect_message(preprocess_data(test_df[1:3, ], "outcome", remove_var = NULL, collapse_corr_feats = FALSE)),
[…]
