### 2020-12-15 41843e7
Describe prefilter_threshold option in news
