+++ NAMESPACE
@@ -23,6 +23,7 @@ export(plot_hp_performance)
 export(plot_model_performance)
 export(preprocess_data)
 export(randomize_feature_order)
+export(remove_singleton_columns)
 export(run_ml)
 export(tidy_perf_data)
 export(train_model)
