### 2020-12-15 89ab4ee
Add @courtneyarmour to contributors
