### 2020-12-15 bb568c1
🎨 Style R code
