+++ tests/testthat/test-preprocess_data.R
@@ -19,9 +19,12 @@ test_df <- data.frame(
 
 test_that("preprocess_data works", {
   expect_equal(
-    expect_message(preprocess_data(test_df, "outcome",
-                                   prefilter_threshold = -1),
-                   "Removed "),
+    expect_message(
+      preprocess_data(test_df, "outcome",
+        prefilter_threshold = -1
+      ),
+      "Removed "
+    ),
     list(dat_transformed = structure(list(
       outcome = c(
         "normal", "normal",
@@ -40,21 +43,24 @@ test_that("preprocess_data works", {
       "var2_a"
     )), removed_feats = c("var5", "var6", "var11"))
   )
-  expect_equal(expect_message(preprocess_data(test_df, "outcome",
-                                              prefilter_threshold = -1,
-                                              group_neg_corr = FALSE)),
-               list(dat_transformed = structure(list(
-    outcome = c(
-      "normal", "normal",
-      "cancer"
-    ), var8 = c(-0.707106781186547, 0.707106781186547, 0),
-    var7_1 = c(1, 1, 0), grp1 = c(-1, 0, 1), grp2 = c(0, 1, 0), grp3 = c(0, 0, 1), grp4 = c(1, 0, 0)
-  ), row.names = c(
-    NA,
-    -3L
-  ), class = c("tbl_df", "tbl", "data.frame")), grp_feats = list(
-    var8 = "var8", var7_1 = "var7_1", grp1 = c("var12", "var1"), grp2 = c("var4_1", "var3_yes", "var2_b", "var9_x", "var10_0"), grp3 = c("var9_y", "var2_c"), grp4 = c("var10_1", "var2_a")
-  ), removed_feats = c("var5", "var6", "var11")))
+  expect_equal(
+    expect_message(preprocess_data(test_df, "outcome",
+      prefilter_threshold = -1,
+      group_neg_corr = FALSE
+    )),
+    list(dat_transformed = structure(list(
+      outcome = c(
+        "normal", "normal",
+        "cancer"
+      ), var8 = c(-0.707106781186547, 0.707106781186547, 0),
+      var7_1 = c(1, 1, 0), grp1 = c(-1, 0, 1), grp2 = c(0, 1, 0), grp3 = c(0, 0, 1), grp4 = c(1, 0, 0)
+    ), row.names = c(
+      NA,
+      -3L
+    ), class = c("tbl_df", "tbl", "data.frame")), grp_feats = list(
+      var8 = "var8", var7_1 = "var7_1", grp1 = c("var12", "var1"), grp2 = c("var4_1", "var3_yes", "var2_b", "var9_x", "var10_0"), grp3 = c("var9_y", "var2_c"), grp4 = c("var10_1", "var2_a")
+    ), removed_feats = c("var5", "var6", "var11"))
+  )
   expect_equal(
     preprocess_data(test_df[1:3, c("outcome", "var1")], "outcome"),
     list(
@@ -92,7 +98,8 @@ test_that("preprocess_data works", {
   )
   expect_equal(
     preprocess_data(test_df[1:3, c("outcome", "var4")], "outcome",
-                    prefilter_threshold = -1),
+      prefilter_threshold = -1
+    ),
     list(
       dat_transformed = dplyr::tibble(
         outcome = c("normal", "normal", "cancer"),
@@ -104,8 +111,9 @@ test_that("preprocess_data works", {
   )
   expect_equal(
     expect_message(preprocess_data(test_df[1:3, ], "outcome",
-                                   method = NULL,
-                                   prefilter_threshold = -1)),
+      method = NULL,
+      prefilter_threshold = -1
+    )),
     list(dat_transformed = structure(list(outcome = c(
       "normal", "normal",
       "cancer"
@@ -131,9 +139,10 @@ test_that("preprocess_data works", {
[…]
