+++ README.md
@@ -1,48 +1,34 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-
-# mikropml
+mikropml
+========
 
 > meek-ROPE em el
 
 User-Friendly R Package for Supervised Machine Learning Pipelines
 
 <!-- badges: start -->
+[![check](https://github.com/SchlossLab/mikropml/workflows/check/badge.svg)](https://github.com/SchlossLab/mikropml/actions?query=workflow%3Acheck+branch%3Amaster) [![codecov](https://codecov.io/gh/SchlossLab/mikropml/branch/master/graph/badge.svg)](https://codecov.io/gh/SchlossLab/mikropml) [![license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/SchlossLab/mikropml/blob/master/LICENSE.md) [![CRAN](https://img.shields.io/cran/v/mikropml?color=blue&label=CRAN&logo=R)](https://CRAN.R-project.org/package=mikropml) [![Conda](https://img.shields.io/conda/vn/conda-forge/r-mikropml)](https://anaconda.org/conda-forge/r-mikropml) [![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/mikropml)](https://cranlogs.r-pkg.org/badges/grand-total/mikropml) <!-- badges: end -->
 
-[![check](https://github.com/SchlossLab/mikropml/workflows/check/badge.svg)](https://github.com/SchlossLab/mikropml/actions?query=workflow%3Acheck+branch%3Amaster)
-[![codecov](https://codecov.io/gh/SchlossLab/mikropml/branch/master/graph/badge.svg)](https://codecov.io/gh/SchlossLab/mikropml)
-[![license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/SchlossLab/mikropml/blob/master/LICENSE.md)
-[![CRAN](https://img.shields.io/cran/v/mikropml?color=blue&label=CRAN&logo=R)](https://CRAN.R-project.org/package=mikropml)
-[![Conda](https://img.shields.io/conda/vn/conda-forge/r-mikropml)](https://anaconda.org/conda-forge/r-mikropml)
-<!-- badges: end -->
+An interface to build machine learning models for classification and regression problems. `mikropml` implements the ML pipeline described by [Topçuoğlu *et al.* (2020)](https://doi.org/doi:10.1128/mBio.00434-20) with reasonable default options for data preprocessing, hyperparameter tuning, cross-validation, testing, model evaluation, and interpretation steps. See the [website](http://www.schlosslab.org/mikropml/) for more information, documentation, and examples.
 
-An interface to build machine learning models for classification and
-regression problems. `mikropml` implements the ML pipeline described by
-[Topçuoğlu *et al.* (2020)](https://doi.org/doi:10.1128/mBio.00434-20)
-with reasonable default options for data preprocessing, hyperparameter
-tuning, cross-validation, testing, model evaluation, and interpretation
-steps. See the [website](http://www.schlosslab.org/mikropml/) for more
-information, documentation, and examples.
+Installation
+------------
 
-## Installation
-
-You can install the latest release from
-[CRAN](https://cran.r-project.org/package=mikropml):
+You can install the latest release from [CRAN](https://cran.r-project.org/package=mikropml):
 
 ``` r
 install.packages('mikropml')
 ```
 
-or the development version from
-[GitHub](https://github.com/SchlossLab/mikRopML):
+or the development version from [GitHub](https://github.com/SchlossLab/mikRopML):
 
 ``` r
 # install.packages("devtools")
 devtools::install_github("SchlossLab/mikropml")
 ```
 
-or install from a terminal using
-[conda](https://docs.conda.io/projects/conda/en/latest/index.html):
[…]
