+++ vignettes/paper.Rmd
@@ -3,6 +3,8 @@ title: "mikropml: User-Friendly R Package for Supervised Machine Learning Pipeli
 output: 
   rmarkdown::html_vignette:
     keep_md: true
+  pdf_document:
+    keep_tex: true
 tags:
   - R
   - machine learning
@@ -44,7 +46,7 @@ affiliations:
     index: 4
   - name: Department of Internal Medicine/Division of Infectious Diseases, University of Michigan
     index: 5
-date: 2020
+date: 2021-01-22
 bibliography: paper.bib
 vignette: >
   %\VignetteIndexEntry{mikropml paper}
