### 2021-01-30 243ff42
Revert "Generate tex for arXiv"
because the vignette fails to render.

This reverts commit 385a7c76d4567bc59ab91c0c29fc74b8824cf05e.
