### 2021-02-27 #2886 @arfon
:wave: @JonnyTran @FedericoComoglio @JDRomano2 – would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission we'd ask you to review here is [_mikropml: User-Friendly R Package for Supervised Machine Learning Pipelines_](https://github.com/SchlossLab/mikropml)
