### 2021-03-01 #2886 @FedericoComoglio
Hi @arfon, I will be happy to help. This would be my first review for JOSS - I will make sure to read the guidelines in the meantime.
