### 2021-03-01 #3073 @whedon
[REVIEW]: mikropml: User-Friendly R Package for Supervised Machine Learning Pipelines
**Submitting author:** @pschloss (<a href="http://orcid.org/0000-0002-6935-4275">Patrick D Schloss</a>)
**Repository:** <a href="https://github.com/SchlossLab/mikropml" target ="_blank">https://github.com/SchlossLab/mikropml</a>
**Version:** v0.0.1
