### 2021-03-19 e5fc674
Edit intro vignette grammar/typos (resolves #231)
