+++ vignettes/introduction.Rmd
@@ -19,8 +19,9 @@ knitr::opts_chunk$set(
 The goal of `mikropml` is to make supervised machine learning (ML) easy for you
 to run while implementing good practices for machine learning pipelines. All you
 need to run ML is one function: `run_ml()`. We've selected sensible default
-arguments related to good practices [@topcuoglu_framework_2020;
-@tang_democratizing_2020], but we allow you to change those arguments to tailor
+arguments related to good practices 
+[@topcuoglu_framework_2020; @tang_democratizing_2020], 
+but we allow you to change those arguments to tailor
 `run_ml()` to the needs of your data.
 
 This document takes you through all of the `run_ml()` inputs,
@@ -75,7 +76,9 @@ but for the purposes of this vignette we'll stick with this example dataset so e
 
 ## The methods we support
 
-All of the methods we use are supported by a great ML wrapper package [`caret`](https://topepo.github.io/caret/), which we use to train our machine learning models.
+All of the methods we use are supported by a great ML wrapper package
+[`caret`](https://topepo.github.io/caret/), which we use to train our machine
+learning models.
 
 The methods we have tested (and their backend packages) are:
 
@@ -86,19 +89,23 @@ The methods we have tested (and their backend packages) are:
 - xgboost (`"xgbTree"`)
 
 For documentation on these methods, as well as many others, you can look at the
-[available models](https://topepo.github.io/caret/available-models.html)
-(or see [here](https://topepo.github.io/caret/train-models-by-tag.html) for a list by tag).
-While we have not vetted the other models used by `caret`, our function is general enough that others might work.
-While we can't promise that we can help with other models,
-feel free to [open an issue on GitHub](https://github.com/SchlossLab/mikropml/issues) if you have questions about other models and we _might_ be able to help.
+[available models](https://topepo.github.io/caret/available-models.html) (or see
+[here](https://topepo.github.io/caret/train-models-by-tag.html) for a list by
+tag). While we have not vetted the other models used by `caret`, our function is
+general enough that others might work. While we can't promise that we can help
+with other models, feel free to [open an issue on
+GitHub](https://github.com/SchlossLab/mikropml/issues) if you have questions
+about other models and we _might_ be able to help.
 
-We will focus on `glmnet` which is our default implementation of L2-regularized logistic regression here, then cover a few other examples towards the end.
+We will first focus on `glmnet`, which is our default implementation of
+L2-regularized logistic regression. Then we will cover a few other examples
+towards the end.
 
 # Before running ML
 
-Before you run ML, you should consider preprocessing your data,
-either on your own or with the `preprocess_data()` function.
-You can learn more about this in the preprocessing vignette: `vignette("preprocess")`.
+Before you execute `run_ml()`, you should consider preprocessing your data,
+either on your own or with the `preprocess_data()` function. You can learn more
+about this in the preprocessing vignette: `vignette("preprocess")`.
 
 # The simplest way to `run_ml()`
 
@@ -109,7 +116,8 @@ You may also want to provide:
 - The outcome column name. By default `run_ml()` will pick the first column, but it's best practice to specify the column name explicitly.
 - A seed so that the results will be reproducible, and so that you get the same results as those you see here (i.e have the same train/test split).
 
-Say we want to use `glmnet`.Then, run ML with:
+Say we want to use logistic regression, then the method we will use is `glmnet`.
+To do so, run ML with:
 
 ```{r, eval = FALSE}
 results <- run_ml(otu_mini_bin,
@@ -143,45 +151,52 @@ because you can learn more from the `caret::train()` documentation.
 names(results$trained_model)
 ```
 
-`test_data` is the partition of the dataset that was used for testing.
[…]
