### 2021-03-19 95ad512
Update NEWS.md
