### 2021-04-09 #3073 @JonnyTran
Hello guys,

I'm about 80% finished with my review process, and below are some of the comments as I've gone through using the `mikropml` package.

---
*Review Comments*
When running the Introduction vignettes on Ubuntu OS, it seems everything works as expected without bugs, which shows it is a well-tested, quality piece of software.

When testing `preprocess_data()` and `run_ml()` on a larger dataset with 1917 rows × 14426 columns, it would take an indeterminately long time. It should have a progress bar indicating the iteration #, and the estimated duration for each iteration,  similar to Python's `tqdm`. Additionally, it can also provide live updates of the loss and other metrics as the model is being trained, to help the user stop the run if it is not satisfactory.

The `run_ml()` function currently implements 5 off-the-shelve ML algorithms, while providing 12 other parameters for training criteria, hyperparameters, feature importance, etc.. If in the future it would support more algorithms, it may introduce too many irrelevant parameters to any certain algorithm, making it difficult to manage. I'd suggest the function to take in 3 objects, e.g. `run_ml(dataset, model, metrics, [args])`, where model hyperparameters are self-contained in an object, and a `metrics` object that allows the user to implement their own functions given the model output and true labels.

---
I'm a heavy Python user who have extensively used scikit-learn and Tensorflow/Pytorch, so these comments were motivated from observing the design choices of those Python ML frameworks. If any of these comments are actionable, I'll also be double posting these as issues to the package github page.

Overall, I fell that the package seems very useful as an introduction to supervised machine learning, and I can see it would have a broad range of applications for exploratory model analysis in various disciplines.

Review time: 4 hours
