### 2021-04-12 dbb4373
Support custom feature groups (resolves #245)
For permutation importance
