+++ R/feature_importance.R
@@ -6,6 +6,10 @@
 #' @param train_data Training data: dataframe of outcome and features.
 #' @inheritParams run_ml
 #' @inheritParams calc_perf_metrics
+#' @param groups Vector of feature names to group together during permutation.
+#'   Each element should be a string with feature names separated by a pipe character (`|`).
+#'   If this is `NULL` (default), correlated features will be grouped together
+#'   based on `corr_thresh`.
 #'
 #' @return Dataframe with performance metrics for when each feature (or group of
 #'   correlated features; `names`) is permuted (`perf_metric`), and differences
@@ -24,6 +28,23 @@
 #'   multiClassSummary, "AUC",
 #'   class_probs = TRUE, method = "glmnet"
 #' )
+#'
+#' # optionally, you can group features together with a custom grouping
+#' get_feature_importance(results$trained_model,
+#'   results$trained_model$trainingData, results$test_data,
+#'   "dx",
+#'   multiClassSummary, "AUC",
+#'   class_probs = TRUE, method = "glmnet",
+#'   groups = c("Otu00007", "Otu00008", "Otu00009", "Otu00011", "Otu00012",
+#'   "Otu00015", "Otu00016", "Otu00018", "Otu00019", "Otu00020", "Otu00022",
+#'   "Otu00023", "Otu00025", "Otu00028", "Otu00029", "Otu00030", "Otu00035",
+#'   "Otu00036", "Otu00037", "Otu00038", "Otu00039", "Otu00040", "Otu00047",
+#'   "Otu00050", "Otu00052", "Otu00054", "Otu00055", "Otu00056", "Otu00060",
+#'   "Otu00003|Otu00002|Otu00005|Otu00024|Otu00032|Otu00041|Otu00053|Otu00014|Otu00021|Otu00017|Otu00031|Otu00057",
+#'   "Otu00013|Otu00006", "Otu00026|Otu00001|Otu00034|Otu00048", "Otu00033|Otu00010",
+#'   "Otu00042|Otu00004", "Otu00043|Otu00027|Otu00049", "Otu00051|Otu00045",
+#'   "Otu00058|Otu00044", "Otu00059|Otu00046")
+#' )
 #' }
 #'
 #' @export
@@ -32,7 +53,7 @@
 get_feature_importance <- function(trained_model, train_data, test_data,
                                    outcome_colname, perf_metric_function,
                                    perf_metric_name, class_probs, method,
-                                   seed = NA, corr_thresh = 1) {
+                                   seed = NA, corr_thresh = 1, groups = NULL) {
   abort_packages_not_installed("future.apply")
 
   # get outcome and features
@@ -40,19 +61,17 @@ get_feature_importance <- function(trained_model, train_data, test_data,
   outcome <- split_dat$outcome
   features <- split_dat$features
 
-  corr_mat <- get_corr_feats(features, corr_thresh = corr_thresh)
-  corr_mat <- dplyr::select_if(corr_mat, !(names(corr_mat) %in% c("corr")))
+  if (is.null(groups)) {
+    groups <- group_correlated_features(features, corr_thresh)
+  }
 
-  grps <- group_correlated_features(corr_mat, features)
+  test_perf_value <- calc_perf_metrics(test_data,
+                                       trained_model,
+                                       outcome_colname,
+                                       perf_metric_function,
+                                       class_probs)[[perf_metric_name]]
 
-  test_perf_value <- calc_perf_metrics(
-    test_data,
-    trained_model,
-    outcome_colname,
-    perf_metric_function,
-    class_probs
-  )[[perf_metric_name]]
-  imps <- future.apply::future_lapply(grps, function(feat) {
+  imps <- future.apply::future_lapply(groups, function(feat) {
[…]
