### 2021-04-12 c4c4797
Point users to get_feature_importance() docs
Resolves #235
