### 2021-04-12 43fe1af
Update NEWS.md
