### 2021-04-12 e3e06bc
Replace spaces in outcome lables with underscores
Resolves #244
