### 2021-04-12 7e1f565
Update NEWS.md
