+++ man/get_performance_tbl.Rd
@@ -58,7 +58,6 @@ A one-row tibble with columns \code{cv_auroc}, column for each of the performanc
 Get model performance metrics as a one-row tibble
 }
 \examples{
-
 \dontrun{
 results <- run_ml(otu_small, "glmnet", kfold = 2, cv_times = 2)
 names(results$trained_model$trainingData)[1] <- "dx"
