### 2021-04-13 f2dbbf9
Describe replace_spaces() return value
