+++ NAMESPACE
@@ -24,6 +24,7 @@ export(plot_model_performance)
 export(preprocess_data)
 export(randomize_feature_order)
 export(remove_singleton_columns)
+export(replace_spaces)
 export(run_ml)
 export(tidy_perf_data)
 export(train_model)
