### 2021-04-13 966a44a
Reflow roxygen doc to 80 chars wide
