+++ R/feature_importance.R
@@ -41,8 +41,10 @@
 #'     "Otu00023", "Otu00025", "Otu00028", "Otu00029", "Otu00030", "Otu00035",
 #'     "Otu00036", "Otu00037", "Otu00038", "Otu00039", "Otu00040", "Otu00047",
 #'     "Otu00050", "Otu00052", "Otu00054", "Otu00055", "Otu00056", "Otu00060",
-#'     "Otu00003|Otu00002|Otu00005|Otu00024|Otu00032|Otu00041|Otu00053|Otu00014|Otu00021|Otu00017|Otu00031|Otu00057",
-#'     "Otu00013|Otu00006", "Otu00026|Otu00001|Otu00034|Otu00048", "Otu00033|Otu00010",
+#'     "Otu00003|Otu00002|Otu00005|Otu00024|Otu00032|Otu00041|Otu00053",
+#'     "Otu00014|Otu00021|Otu00017|Otu00031|Otu00057",
+#'     "Otu00013|Otu00006", "Otu00026|Otu00001|Otu00034|Otu00048",
+#'     "Otu00033|Otu00010",
 #'     "Otu00042|Otu00004", "Otu00043|Otu00027|Otu00049", "Otu00051|Otu00045",
 #'     "Otu00058|Otu00044", "Otu00059|Otu00046"
 #'   )
