### 2021-04-14 SchlossLab/mikropml#244 @kelly-sovacool
Hi @JonnyTran, we went with your suggestion for `preprocess_data()` to automatically prepare the outcome column. Thanks for spotting this!
