+++ .github/pull_request_template.md
@@ -9,6 +9,7 @@
 
 (~Strikethrough~ any points that are not applicable.)
 
+- [ ] Write unit tests for any new functionality.
 - [ ] Update roxygen comments & vignettes if there are any API changes.
 - [ ] Update `NEWS.md` if this includes any user-facing changes. 
 - [ ] The check workflow succeeds on your most recent commit.
