+++ README.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 )
 ```
 
-# mikropml
+# mikropml <a href='http://www.schlosslab.org/mikropml'><img src='man/figures/logo.png' align="right" height="120" /></a>
 
 > meek-ROPE em el
 
@@ -106,6 +106,12 @@ If you'd like to contribute, see our guidelines
 Please note that the mikropml project is released with a [Contributor Code of
 Conduct](http://www.schlosslab.org/mikropml/CODE_OF_CONDUCT.html). By
 contributing to this project, you agree to abide by its terms.
+
+## License
+
+The mikropml package is licensed under [the MIT license](license.md).
+The mikropml logo is licensed under the 
+[CC BY 4.0 license](https://creativecommons.org/licenses/by/4.0/).
 
 ## Why the name?
 
