### 2021-04-28 3e7f22e
Add the logo to the paper
