### 2021-04-28 3475f43
Remove hyperlink for downloads badge
