### 2021-04-28 #3073 @FedericoComoglio
Hi everyone,

I completed my review of the `mikropml` package. I would like to apologize once again with the authors for the delay.

This package offers off-the-shelf ML pipelines for regression and classification tasks. All methods are supported by the `caret` package (in fact, most functions offer a sensible wrapper around `caret` functions). The examples are informative and well-crafted. The vignettes describing data preprocessing and hyperparameter tuning offer sufficient context for a user with limited exposure to ML. While most of the functionalities implemented by the package are covered by alternative frameworks, most notably the [tidymodels](https://www.tidymodels.org/) ecosystem, `mikropml` largely facilitates execution of standard ML pipelines. I extensively tested the `run_ml()` function with different combinations of parameters and hyperparameters on a small (the `otu_mini_bin` dataset that is shipped with the package) as well as a larger (~10k x 15) dataset. I am positively impressed about the package functionalities and the overall quality of its implementation and documentation.

## Major

- None.

## Minor

- I agree with @JonnyTran 's suggestion to display time to completion estimates and relevant metrics (e.g. loss or chosen metrics for current iteration) during training. The average user will find this helpful. This would also be the case when parallelizing `run_ml()` execution.
- In a number of places across the vignette, "machine learning" (e.g. "run machine learning" in the "Preprocessing data" vignette) is used in place of a likely more appropriate "model training"/"train model"
- For training random forest models, it is unfortunate that the number of trees can not be tuned. However, this is a consequence of design choices that pertain to the `caret` package and hence not directly related to this work.

## Conclusion

I am convinced `mikropml` would be of potential interest to users looking for ML pipelines to solve common regression and classification problems. In particular, users with little or no prior exposure to ML would find sufficient guidance in the package documentation and, most importantly, a safety net provided by the implementation of several best practices.

Review time: 3+ hours
