### 2021-04-29 SchlossLab/mikropml#249 @kelly-sovacool
Implement progress bar for long-running functions
From https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-817023969:

> When testing preprocess_data() and run_ml() on a larger dataset with 1917 rows × 14426 columns, it would take an indeterminately long time. It should have a progress bar indicating the iteration #, and the estimated duration for each iteration, similar to Python's tqdm.

And https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-828720647:

> I agree with @JonnyTran 's suggestion to display time to completion estimates and relevant metrics (e.g. loss or chosen metrics for current iteration) during training. The average user will find this helpful. This would also be the case when parallelizing `run_ml()` execution.

I'll probably use https://github.com/r-lib/progress and make it optional.
