### 2021-04-29 SchlossLab/mikropml#250 @kelly-sovacool
Report live updates of metrics during training
From https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-817023969:

> When testing preprocess_data() and run_ml() on a larger dataset with 1917 rows × 14426 columns, it would take an indeterminately long time. It should have a progress bar indicating the iteration #, and the estimated duration for each iteration, similar to Python's tqdm. Additionally, it can also provide live updates of the loss and other metrics as the model is being trained, to help the user stop the run if it is not satisfactory.

And https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-828720647:

> I agree with @JonnyTran 's suggestion to display time to completion estimates and relevant metrics (e.g. loss or chosen metrics for current iteration) during training. The average user will find this helpful. This would also be the case when parallelizing `run_ml()` execution.
