### 2021-04-29 SchlossLab/mikropml#251 @kelly-sovacool
Refactor run_ml() API for future expandability
From https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-817023969:

> The `run_ml()` function currently implements 5 off-the-shelve ML algorithms, while providing 12 other parameters for training criteria, hyperparameters, feature importance, etc.. If in the future it would support more algorithms, custom metrics, or training parameters, I'd imagine there'll be limitations imposed by the function arguments. I'd suggest the function to take in 3 objects, e.g. `run_ml(dataset, model, metrics, [args])`, where a metrics object can allow the user select standard metrics or define their own metric functions given the model output and true labels.
