### 2021-04-29 SchlossLab/mikropml#252 @kelly-sovacool
Test multi-outcome datasets
One of our JOSS reviewers stated that `run_ml()` "doesn't seem to be able to support multi-label classifications." (https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-817023969)

We include a multi-outcome dataset (`otu_mini_multi`) in the package and results from running it with glmnet (`otu_mini_multi_results_glmnet`), but I just noticed that we never wrote any unit tests for it!
