### 2021-04-29 SchlossLab/mikropml#253 @kelly-sovacool
Use the phrase "model training" rather than "ML" where appropriate
From https://github.com/openjournals/joss-reviews/issues/3073#issuecomment-828720647:

> In a number of places across the vignette, "machine learning" (e.g. "run machine learning" in the "Preprocessing data" vignette) is used in place of a likely more appropriate "model training"/"train model"
