### 2021-04-30 SchlossLab/mikropml#252 @zenalapp
Multi-label classification is when you can assign multiple labels to each sample/instance. I don't think `caret` allows this functionality so I think it's probably out of the scope of this package. We can add it as a limitation to the manuscript and vignette. Thoughts @kelly-sovacool @BTopcuoglu?

I can add in a test for a multi-outcome dataset either way, but I don't think this is what the reviewer was referencing.
