### 2021-04-30 SchlossLab/mikropml#252 @kelly-sovacool
Ohhh I see, I confused the similar terminology.

Let's note the limitation in one of the vignettes, maybe the intro one? I don't know that we need to add it to the paper since it's really out of scope and not even caret or tidymodels support it.

But yes either way we do actually need a unit test for the multi-outcome dataset.
