### 2021-04-30 SchlossLab/mikropml#250 @zenalapp
I looked at the documentation for `caret::train`, which is the function we use to train the model. Unfortunately I don't see an option to display updates on relevant metrics as the model is being trained :( So not sure there's a way for us to easily implement this.

https://rdrr.io/cran/caret/man/train.html
