### 2021-04-30 3fdd11a
Add test for multi-outcome labels resolves #252
