+++ tests/testthat/test-run_ml.R
@@ -151,3 +151,19 @@ test_that("run_ml errors if outcome_colname not in dataframe", {
     "Outcome 'not_a_colname' not in column names of data."
   )
 })
+
+test_that("run_ml works for multiclass outcome", {
+  skip_on_cran()
+  expect_equal_ml_results(
+    expect_warning(expect_message(run_ml(otu_mini_multi, # use built-in hyperparameters
+                                         "glmnet",
+                                         outcome_colname = "dx",
+                                         find_feature_importance = TRUE,
+                                         seed = 2019,
+                                         cv_times = 2,
+                                         group = otu_mini_multi_group
+    ), "Using 'dx' as the outcome column"),"`caret::train\\(\\)` issued the following warning:"),
+    otu_mini_multi_results_glmnet
+  )
+})
+
