### 2021-04-30 SchlossLab/mikropml#251 @zenalapp
`run_ml()` already takes these objects:
1. `dataset`: The input dataset.
2. `method`: The ML model to be used. While we only officially support 5 models, all of the models supported by `caret` (https://topepo.github.io/caret/available-models.html) should work in our package. If `caret` supports additional models in the future, these should also work in `mikropml`. We realize that the model options are not as generalizable as e.g. PyTorch, since users must choose from options supported by `caret`. However, our code heavily relies on `caret` to perform the underlying model training. Additionally, as `mikropml` is oriented toward beginner practitioners, we believe that it does not need to provide the option to include custom models.
3. `perf_metric_function` and `perf_metric_name`: The performance metric to be used. We chose sensible defaults, but the user can provide their own performance metrics if they would like.
4. `hyperparameters`: The values of hyperparameters in the model that the user would like to tune.
