### 2021-04-30 63df4d0
Update NEWS.md
Co-authored-by: Kelly Sovacool <kelly-sovacool@users.noreply.github.com>
