+++ README.md
@@ -14,7 +14,7 @@ User-Friendly R Package for Supervised Machine Learning Pipelines
 [![license](https://img.shields.io/badge/license-MIT-blue.svg)](https://github.com/SchlossLab/mikropml/blob/master/LICENSE.md)
 [![CRAN](https://img.shields.io/cran/v/mikropml?color=blue&label=CRAN&logo=R)](https://CRAN.R-project.org/package=mikropml)
 [![Conda](https://img.shields.io/conda/vn/conda-forge/r-mikropml)](https://anaconda.org/conda-forge/r-mikropml)
-[![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/mikropml)](https://cranlogs.r-pkg.org/badges/grand-total/mikropml)
+<https://cranlogs.r-pkg.org/badges/grand-total/mikropml>
 <!-- badges: end -->
 
 An interface to build machine learning models for classification and
@@ -51,9 +51,9 @@ conda install -c conda-forge r-mikropml
 
 ### Dependencies
 
--   Imports: caret, dplyr, e1071, glmnet, kernlab, MLmetrics,
+  - Imports: caret, dplyr, e1071, glmnet, kernlab, MLmetrics,
     randomForest, rlang, rpart, stats, utils, xgboost
--   Suggests: doFuture, foreach, future, future.apply, ggplot2, knitr,
+  - Suggests: doFuture, foreach, future, future.apply, ggplot2, knitr,
     purrr, rmarkdown, testthat, tidyr
 
 ## Usage
@@ -91,4 +91,4 @@ This package was originally implemented as a machine learning pipeline
 for microbiome-based classification problems (see [Topçuoğlu *et al.*
 2020](https://doi.org/10.1128/mBio.00434-20)). We realized that these
 methods are applicable in many other fields too, but stuck with the name
-because we like it!
+because we like it\!
