### 2021-05-03 9537825
Remove unnecessary comments
Co-authored-by: Kelly Sovacool <kelly-sovacool@users.noreply.github.com>
