+++ .github/pull_request_template.md
@@ -12,4 +12,4 @@
 - [ ] Write unit tests for any new functionality.
 - [ ] Update roxygen comments & vignettes if there are any API changes.
 - [ ] Update `NEWS.md` if this includes any user-facing changes. 
-- [ ] The check workflow succeeds on your most recent commit.
+- [ ] The check workflow succeeds on your most recent commit. **This is always required before the PR can be merged.**
