### 2021-05-04 97708f2
Support optional progress bar for preprocess_data()
