### 2021-05-04 SchlossLab/mikropml#249 @kelly-sovacool
Unfortunately the progress bar is nearly useless for `run_ml()` since the bulk of the execution time is spent running `caret::train()`. We won't be able to fix that without writing our own train function, which is probably not worth it.

The progress bar is more informative for `preprocess_data()`, so I implemented it there.
