### 2021-05-04 SchlossLab/mikropml#249 @JonnyTran
Wonderful @kelly-sovacool ! I was wondering if it's also possible to implement a progress bar for the feature importance, since computing it can take a long time?
