+++ R/preprocess_data.R
@@ -49,13 +49,13 @@ preprocess_data <- function(dataset, outcome_colname,
                             to_numeric = TRUE, group_neg_corr = TRUE,
                             prefilter_threshold = 1) {
 
-  if (isTRUE(check_packages_installed('progress'))) {
-    progbar <- progress::progress_bar$new(format = "preprocess_data() :bar :percent | elapsed: :elapsed | eta: :eta",
-                                          total = 9)
-    progbar$tick(0)
-  } else {
-    progbar <- NULL
-  }
+  progbar <- pbinit(total = 19, message = "preprocessing")
+  # progbar <-
+  #   progress::progress_bar$new(
+  #     format = ":current | elapsed: :elapsed | eta: :eta",
+  #     total = 18
+  #   )
+  # progbar$tick(0)
 
   check_dataset(dataset)
   check_outcome_column(dataset, outcome_colname, check_values = FALSE)
@@ -74,9 +74,9 @@ preprocess_data <- function(dataset, outcome_colname,
   }
   pbtick(progbar)
 
-  nv_feats <- process_novar_feats(features)
+  nv_feats <- process_novar_feats(features, progbar = progbar)
   pbtick(progbar)
-  split_feats <- process_cat_feats(nv_feats$var_feats)
+  split_feats <- process_cat_feats(nv_feats$var_feats, progbar = progbar)
   pbtick(progbar)
   cont_feats <- process_cont_feats(split_feats$cont_feats, method)
   pbtick(progbar)
@@ -105,7 +105,9 @@ preprocess_data <- function(dataset, outcome_colname,
       removed_feats <- c(removed_feats, feats$removed)
       pbtick(progbar)
     }
-    feats_and_grps <- collapse_correlated_features(processed_feats, group_neg_corr)
+    feats_and_grps <- collapse_correlated_features(processed_feats,
+                                                   group_neg_corr,
+                                                   progbar = progbar)
     processed_feats <- feats_and_grps$features
     grp_feats <- feats_and_grps$grp_feats
   }
@@ -114,7 +116,6 @@ preprocess_data <- function(dataset, outcome_colname,
   # combine outcome and features
   dat_transformed <- dplyr::bind_cols(split_dat$outcome, processed_feats) %>%
     dplyr::as_tibble()
-  pbtick(progbar)
 
   if (!is.null(progbar)) {progbar$terminate()}
   return(list(
@@ -210,6 +211,7 @@ remove_singleton_columns <- function(dat, threshold = 1) {
 #' Process features with no variation
 #'
 #' @param features dataframe of features for machine learning
+#' @param progbar optional progress bar (default: `NULL`)
 #'
 #' @return list of two dataframes: features with variability (unprocessed) and without (processed)
 #' @noRd
@@ -217,7 +219,7 @@ remove_singleton_columns <- function(dat, threshold = 1) {
 #'
 #' @examples
 #' process_novar_feats(mikropml::otu_small[, 2:ncol(otu_small)])
-process_novar_feats <- function(features) {
+process_novar_feats <- function(features, progbar = NULL) {
   novar_feats <- NULL
   var_feats <- NULL
   if (!is.null(features)) {
@@ -229,7 +231,7 @@ process_novar_feats <- function(features) {
       length(unique(x[!is.na(x)])) == 1
     })
     novar_feats <- features %>% dplyr::select_if(novar_feats_bool)
-
+    pbtick(progbar)
     # change categorical features with no variation to zero
     sapply_fn <- select_apply(fun = "sapply")
     novar_feats <- sapply_fn(novar_feats, function(x) {
@@ -243,7 +245,7 @@ process_novar_feats <- function(features) {
     if (ncol(novar_feats) == 0) {
[…]
