+++ R/run_ml.R
@@ -158,7 +158,7 @@ run_ml <-
     )
 
     model_formula <- stats::as.formula(paste(outcome_colname, "~ ."))
-
+    message('Training the model...')
     trained_model_caret <- train_model(
       model_formula,
       train_data,
