+++ R/preprocess_data.R
@@ -26,8 +26,8 @@
 #' - `grp_feats`: If features were grouped together, a named list of the features corresponding to each group.
 #' - `removed_feats`: Any features that were removed during preprocessing (e.g. because there was zero variance or near-zero variance for those features).
 #'
-#' If the `progress` package is installed, a progress bar with time elapsed
-#' and estimated time to completion will be displayed.
+#' If the `progressr` package is installed, a progress bar with time elapsed
+#' and estimated time to completion can be displayed.
 #'
 #' @section More details:
 #'
@@ -43,23 +43,31 @@
 #'
 #' @examples
 #' preprocess_data(mikropml::otu_small, "dx")
+#'
+#' # the function can show a progress bar if you have the progressr package installed
+#' ## optionally, specify the progress bar format
+#' progressr::handlers(progressr::handler_progress(
+#'         format = ":message :bar :percent | elapsed: :elapsed | eta: :eta",
+#'         clear = FALSE,
+#'         show_after = 0))
+#' ## tell progressor to always report progress
+#' progressr::handlers(global = TRUE)
+#' ## run the function and watch the live progress udpates
+#' dat_preproc <- preprocess_data(mikropml::otu_small, "dx")
 preprocess_data <- function(dataset, outcome_colname,
                             method = c("center", "scale"),
                             remove_var = "nzv", collapse_corr_feats = TRUE,
                             to_numeric = TRUE, group_neg_corr = TRUE,
                             prefilter_threshold = 1) {
 
-  progbar <- pbinit(total = 19, message = "preprocessing")
-  # progbar <-
-  #   progress::progress_bar$new(
-  #     format = ":current | elapsed: :elapsed | eta: :eta",
-  #     total = 18
-  #   )
-  # progbar$tick(0)
-
+  progbar <- NULL
+  if (isTRUE(check_packages_installed('progressr'))) {
+    progbar <- progressr::progressor(steps = 18, message = "preprocessing")
+  }
   check_dataset(dataset)
   check_outcome_column(dataset, outcome_colname, check_values = FALSE)
   check_remove_var(remove_var)
+  pbtick(progbar)
   dataset[[outcome_colname]] <- replace_spaces(dataset[[outcome_colname]])
   dataset <- rm_missing_outcome(dataset, outcome_colname)
   split_dat <- split_outcome_features(dataset, outcome_colname)
@@ -117,7 +125,6 @@ preprocess_data <- function(dataset, outcome_colname,
   dat_transformed <- dplyr::bind_cols(split_dat$outcome, processed_feats) %>%
     dplyr::as_tibble()
 
-  pbend(progbar)
   return(list(
     dat_transformed = dat_transformed,
     grp_feats = grp_feats,
