+++ R/utils.R
@@ -152,77 +152,32 @@ replace_spaces <- function(x, new_char = "_") {
   gsub(" ", new_char, x)
 }
 
-#' Create a progress bar if the `progress` package is installed.
-#'
-#' The progress bar persists after completion.
-#'
-#' @param total total ticks passed to `progress::progress_bar$new()`
-#'   (default: `10`).
-#' @param message an optional message pasted in front of the progress bar
-#'   (default: `""`).
-#'
-#' @return a new `progress::progress_bar`, or `NULL` if the `progress` package
-#'   is not installed
-#' @noRd
-#' @author Kelly Sovacool \email{sovacool@@umich.edu}
-#'
-#' @examples
-#' progbar <- pbinit(total = 5, message = "looping")
-#' for (i in 1:5) {
-#'   pbtick(progbar)
-#'   Sys.sleep(0.5)
-#' }
-pbinit <- function(total = 10, message = "") {
-  pb <- NULL
-  if (isTRUE(check_packages_installed('progress'))) {
-    pb <-
-      progress::progress_bar$new(
-        format = paste(message, ":bar :percent | elapsed: :elapsed | eta: :eta"),
-        total = total,
-        clear = FALSE,
-        show_after = 0
-      )
-    pb$tick(0)
-  }
-  return(pb)
-}
-
 #' Update progress if the progress bar is not `NULL`
 #'
-#' @param pb a progress bar created with `progress::progress_bar$new()`
+#' @param pb a progress bar created with `progressr`.
+#' @param message optional message to report (default: `NULL`).
 #'
 #' @noRd
 #' @author Kelly Sovacool \email{sovacool@@umich.edu}
 #'
 #' @examples
-#' progbar <- pbinit(total = 5)
-#' for (i in 1:5) {
-#'   pbtick(progbar)
-#'   Sys.sleep(0.5)
+#' f <- function () {
+#'   pb <- progressr::progressor(steps = 5, message = 'looping')
+#'   for (i in 1:5) {
+#'     pbtick(pb)
+#'     Sys.sleep(0.5)
+#'   }
 #' }
-pbtick <- function(pb) {
-  if (!is.null(pb)) { pb$tick() }
-  invisible()
-}
-
-#' End the progress bar if it is not `NULL`.
-#'
-#' This is useful to jump to the end of the bar when a function has
-#' finished, but fewer than the estimated total number of ticks were used.
-#'
-#' @param pb a progress bar created with `progress::progress_bar$new()`
-#'
-#' @noRd
-#' @author Kelly Sovacool \email{sovacool@@umich.edu}
-#'
-#' @examples
-#' progbar <- pbinit(total = 10)
-#' for (i in 1:5) {
-#'   pbtick(progbar)
-#'   Sys.sleep(0.5)
-#' }
-#' pbend(progbar)
-pbend <- function(pb) {
-  if (!is.null(pb)) { pb$update(1) }
+#' progressr::with_progress(
+#'   f()
+#' )
+pbtick <- function(pb, message = NULL) {
+  if (!is.null(pb)) {
+    if (!is.null(message)) {
+      pb(message)
+    } else {
+      pb()
+    }
+  }
   invisible()
 }
