### 2021-05-05 SchlossLab/mikropml#249 @kelly-sovacool
Nice idea @JonnyTran, I'll implement it for feature importance too in #257.
