+++ README.md
@@ -54,7 +54,7 @@ conda install -c conda-forge r-mikropml
   - Imports: caret, dplyr, e1071, glmnet, kernlab, MLmetrics,
     randomForest, rlang, rpart, stats, utils, xgboost
   - Suggests: doFuture, foreach, future, future.apply, ggplot2, knitr,
-    purrr, rmarkdown, testthat, tidyr
+    progress, progressr, purrr, rmarkdown, testthat, tidyr
 
 ## Usage
 
