+++ man/preprocess_data.Rd
@@ -69,9 +69,10 @@ preprocess_data(mikropml::otu_small, "dx")
 # the function can show a progress bar if you have the progressr package installed
 ## optionally, specify the progress bar format
 progressr::handlers(progressr::handler_progress(
-        format = ":message :bar :percent | elapsed: :elapsed | eta: :eta",
-        clear = FALSE,
-        show_after = 0))
+  format = ":message :bar :percent | elapsed: :elapsed | eta: :eta",
+  clear = FALSE,
+  show_after = 0
+))
 ## tell progressor to always report progress
 progressr::handlers(global = TRUE)
 ## run the function and watch the live progress udpates
