### 2021-05-06 1169d13
Cleanup leftover debugging code
