### 2021-05-06 f119247
Use nperms arg in get_feature_importance()
