+++ tests/testthat/test-utils.R
@@ -74,7 +74,7 @@ test_that("replace_spaces works", {
 test_that("pbtick() updates the progress bar", {
   f <- function() {
     pb <- progressr::progressor(steps = 5)
-    pbtick(pb)
+    pbtick(pb, message = 'progress!')
   }
   expect_condition(expect_invisible(f()))
 })
