### 2021-05-06 eced9e5
Test that get_outcome_type() errors when expected
