### 2021-05-06 ccc20c9
Fix bug that allowed numeric outcomes of length 1
