### 2021-05-06 bf09557
Reduce repetitive test code
