### 2021-05-06 3ed23c8
Put back pbtick() calls in preprocess_data
