### 2021-05-06 e27e78b
Increase total steps for preprocess progress
Based on Zena's clinical dataset which should enter all the if
statements.
