+++ vignettes/review-responses.md
@@ -1,22 +1,22 @@
 We thank the reviewers for their thoughtful feedback. We have responded to all reviewer comments below. Feel free to let us know if you have additional questions.
 
-## JonnyTran reviews
+## @JonnyTran reviews
 
-> When testing `preprocess_data()` and `run_ml()` on a larger dataset with 1917 rows × 14426 columns, it would take an indeterminately long time. It should have a progress bar indicating the iteration #, and the estimated duration for each iteration, similar to Python's `tqdm`. 
+> When testing `preprocess_data()` and `run_ml()` on a larger dataset with 1917 rows × 14426 columns, it would take an indeterminately long time. It should have a progress bar indicating the iteration #, and the estimated duration for each iteration, similar to Python's `tqdm`.
 
 We opened and resolved issue [#249](https://github.com/SchlossLab/mikropml/issues/249) related to this suggestion.
 We have implemented this feature for `preprocess_data()` and the feature importance part of `run_ml()`.
 Unfortunately, we cannot do it for model training as we use `caret::train()`, which does not support having a progress bar.
- 
+
 > Additionally, it can also provide live updates of the loss and other metrics as the model is being trained, to help the user stop the run if it is not satisfactory.
 
 Another limitation of using `caret::train()` to train the model is that there is no option to display updates on relevant metrics as the model is being trained.
-As much of our package is based on using this function, we unfortunately cannot easily implement this suggestion. 
+As much of our package is based on using this function, we unfortunately cannot easily implement this suggestion.
 (See issue [#250](https://github.com/SchlossLab/mikropml/issues/250))
 
-> I also encountered a bug when running `run_ml()` on a dataset with class name containing spaces in the label column. I can mitigate the bug by replacing all `" "`'s with `"_"`'s. 
+> I also encountered a bug when running `run_ml()` on a dataset with class name containing spaces in the label column. I can mitigate the bug by replacing all `" "`'s with `"_"`'s.
 
-Thank you for bringing this to our attention. We fixed this issue (issue [#244](https://github.com/SchlossLab/mikropml/issues/244)).
+Thank you for bringing this to our attention. We fixed this issue ([#244](https://github.com/SchlossLab/mikropml/issues/244)).
 
 > Currently, it doesn't seem to be able to support multi-label classifications.
 
@@ -24,8 +24,8 @@ Unfortunately `caret` does not support this functionality so we believe it's out
 
 > The `run_ml()` function currently implements 5 off-the-shelve ML algorithms, while providing 12 other parameters for training criteria, hyperparameters, feature importance, etc.. If in the future it would support more algorithms, custom metrics, or training parameters, I'd imagine there'll be limitations imposed by the function arguments. I'd suggest the function to take in 3 objects, e.g. `run_ml(dataset, model, metrics, [args])`, where a metrics object can allow the user select standard metrics or define their own metric functions given the model output and true labels.
 
-This is a very good point that we have responded to in issue [#251](https://github.com/SchlossLab/mikropml/issues/251). 
-We chose to structure `run_ml` in a way that is easier for our users, rather than extremely flexible. 
+This is a very good point that we have responded to in issue [#251](https://github.com/SchlossLab/mikropml/issues/251).
+We chose to structure `run_ml` in a way that is easier for our users, rather than extremely flexible.
 If users want a more customizable experience, especially related to what models they want to use, we expect them to use other packages not developed for machine learning novices.
 Furthermore, `run_ml()` already takes these objects:
 
@@ -34,7 +34,7 @@ Furthermore, `run_ml()` already takes these objects:
[…]
