### 2021-05-07 8b49924
Update metadata for v1.0.0
