### 2021-05-13 d013a2e
Update NEWS for v1.0.0
