+++ README.md
@@ -15,6 +15,7 @@ User-Friendly R Package for Supervised Machine Learning Pipelines
 [![CRAN](https://img.shields.io/cran/v/mikropml?color=blue&label=CRAN&logo=R)](https://CRAN.R-project.org/package=mikropml)
 [![Conda](https://img.shields.io/conda/vn/conda-forge/r-mikropml)](https://anaconda.org/conda-forge/r-mikropml)
 ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/mikropml)
+[![JOSS](https://joss.theoj.org/papers/72bf31a3f51f8fc273ef6b99bd04ede1/status.svg)](https://joss.theoj.org/papers/72bf31a3f51f8fc273ef6b99bd04ede1)
 <!-- badges: end -->
 
 An interface to build machine learning models for classification and
@@ -51,10 +52,10 @@ conda install -c conda-forge r-mikropml
 
 ### Dependencies
 
-  - Imports: caret, dplyr, e1071, glmnet, kernlab, MLmetrics,
+-   Imports: caret, dplyr, e1071, glmnet, kernlab, MLmetrics,
     randomForest, rlang, rpart, stats, utils, xgboost
-  - Suggests: doFuture, foreach, future, future.apply, ggplot2, knitr,
-    progress, progressr, purrr, rmarkdown, testthat, tidyr
+-   Suggests: doFuture, foreach, future, future.apply, ggplot2, knitr,
+    purrr, rmarkdown, testthat, tidyr
 
 ## Usage
 
@@ -98,4 +99,4 @@ This package was originally implemented as a machine learning pipeline
 for microbiome-based classification problems (see [Topçuoğlu *et al.*
 2020](https://doi.org/10.1128/mBio.00434-20)). We realized that these
 methods are applicable in many other fields too, but stuck with the name
-because we like it\!
+because we like it!
