+++ README.Rmd
@@ -28,6 +28,7 @@ User-Friendly R Package for Supervised Machine Learning Pipelines
 [![CRAN](https://img.shields.io/cran/v/mikropml?color=blue&label=CRAN&logo=R)](https://CRAN.R-project.org/package=mikropml)
 [![Conda](https://img.shields.io/conda/vn/conda-forge/r-mikropml)](https://anaconda.org/conda-forge/r-mikropml)
 ![Downloads](https://cranlogs.r-pkg.org/badges/grand-total/mikropml)
+[![JOSS](https://joss.theoj.org/papers/72bf31a3f51f8fc273ef6b99bd04ede1/status.svg)](https://joss.theoj.org/papers/72bf31a3f51f8fc273ef6b99bd04ede1)
 <!-- badges: end -->
 
 An interface to build machine learning models for classification and regression
