### 2021-05-13 814d552
Tweak URLs for CRAN
