+++ README.md
@@ -1,7 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-# mikropml <a href='http://www.schlosslab.org/mikropml'><img src='man/figures/logo.png' align="right" height="120" /></a>
+# mikropml <a href='http://www.schlosslab.org/mikropml/'><img src='man/figures/logo.png' align="right" height="120" /></a>
 
 > meek-ROPE em el
 
@@ -87,7 +87,7 @@ contributing to this project, you agree to abide by its terms.
 
 ## License
 
-The mikropml package is licensed under [the MIT license](license.md).
+The mikropml package is licensed under [the MIT license](LICENSE.md).
 Text and images included in this repository, including the mikropml
 logo, are licensed under the [CC BY 4.0
 license](https://creativecommons.org/licenses/by/4.0/).
