### 2021-05-13 1a0fbe8
Add citation instructions to README
