### 2021-05-14 #3073 @kelly-sovacool
Hi @arfon, the DOI is `10.5281/zenodo.4759346`. I have verified that the title and authors in the archive are the same as those in the JOSS paper.
