### 2021-05-14 #3073 @pschloss
Wonderful - thank you all for the great feed back and the wonderful experience!

Pat

> On May 14, 2021, at 02:04, Arfon Smith ***@***.***> wrote:
>
>
> @JonnyTran <https://github.com/JonnyTran>, @FedericoComoglio <https://github.com/FedericoComoglio> - many thanks for your reviews here! JOSS relies up volunteer effort from folks like you and we simply wouldn't be able to do this without you ✨ .
>
> @pschloss <https://github.com/pschloss> - your paper is now accepted and published in JOSS ⚡🚀💥
>
